eliminate some warnings

+++ src/osgEarthDrivers/engine_osgterrain/SinglePassTerrainTechnique.cpp
@@ -901,19 +901,19 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
     {
       int min_i = (int)floor((*mr)._ndcMin.x() * (double)(numColumns-1));
       if (min_i < 0) min_i = 0;
-      if (min_i >= numColumns) min_i = numColumns - 1;
+      if (min_i >= (int)numColumns) min_i = numColumns - 1;
 
       int min_j = (int)floor((*mr)._ndcMin.y() * (double)(numRows-1));
       if (min_j < 0) min_j = 0;
-      if (min_j >= numColumns) min_j = numColumns - 1;
+      if (min_j >= (int)numColumns) min_j = numColumns - 1;
 
       int max_i = (int)ceil((*mr)._ndcMax.x() * (double)(numColumns-1));
       if (max_i < 0) max_i = 0;
-      if (max_i >= numColumns) max_i = numColumns - 1;
+      if (max_i >= (int)numColumns) max_i = numColumns - 1;
 
       int max_j = (int)ceil((*mr)._ndcMax.y() * (double)(numRows-1));
       if (max_j < 0) max_j = 0;
-      if (max_j >= numColumns) max_j = numColumns - 1;
+      if (max_j >= (int)numColumns) max_j = numColumns - 1;
 
       if (min_i >= 0 && max_i >= 0 && min_j >= 0 && max_j >= 0)
       {
@@ -1267,7 +1267,7 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
             ss_verts->reserve(ss_verts->size() + outVerts->size() * 4 + skirtIndices.size() * 2);
 
             //Add a primative set for each continuous skirt strip
-            for (int p=0; p < skirtIndices.size(); p++)
+            for (int p=0; p < (int)skirtIndices.size(); p++)
             {
               int cursor = ss_verts->size();
 
@@ -1375,7 +1375,7 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
 
             if (orig_i < 0)
             {
-              if (skirtBreaks.back() != skirtVerts->size())
+              if (skirtBreaks.back() != (int)skirtVerts->size())
                 skirtBreaks.push_back(skirtVerts->size());
             }
             else
@@ -1406,7 +1406,7 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
             int orig_i = indices[r*numColumns+(numColumns-1)];
             if (orig_i < 0)
             {
-              if (skirtBreaks.back() != skirtVerts->size())
+              if (skirtBreaks.back() != (int)skirtVerts->size())
                 skirtBreaks.push_back(skirtVerts->size());
             }
             else
@@ -1437,7 +1437,7 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
             int orig_i = indices[(numRows-1)*numColumns+c];
             if (orig_i < 0)
             {
-              if (skirtBreaks.back() != skirtVerts->size())
+              if (skirtBreaks.back() != (int)skirtVerts->size())
                 skirtBreaks.push_back(skirtVerts->size());
             }
             else
@@ -1468,7 +1468,7 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
             int orig_i = indices[r*numColumns];
             if (orig_i < 0)
             {
-              if (skirtBreaks.back() != skirtVerts->size())
+              if (skirtBreaks.back() != (int)skirtVerts->size())
                 skirtBreaks.push_back(skirtVerts->size());
             }
             else
@@ -1497,7 +1497,7 @@ SinglePassTerrainTechnique::createGeometry( const TileFrame& tilef )
 
         //Add a primative set for each continuous skirt strip
         skirtBreaks.push_back(skirtVerts->size());
-        for (int p=1; p < skirtBreaks.size(); p++)
+        for (int p=1; p < (int)skirtBreaks.size(); p++)
           skirt->addPrimitiveSet( new osg::DrawArrays( GL_TRIANGLE_STRIP, skirtBreaks[p-1], skirtBreaks[p] - skirtBreaks[p-1] ) );
     }
